chrom : fix cargo clippy error in runtime
