+++ kclvm/runtime/src/stdlib/assert_api.rs
@@ -7,10 +7,7 @@ type kclvm_value_ref_t = ValueRef;
 
 #[no_mangle]
 #[runtime_fn]
-pub unsafe extern "C" fn kclvm_assert(
-    value: *const kclvm_value_ref_t,
-    msg: *const kclvm_value_ref_t,
-) {
+pub extern "C" fn kclvm_assert(value: *const kclvm_value_ref_t, msg: *const kclvm_value_ref_t) {
     let value = ptr_as_ref(value);
     let msg = ptr_as_ref(msg);
