+++ kclvm/src/lib.rs
@@ -53,7 +53,8 @@ pub unsafe extern "C" fn kclvm_cli_run(args: *const i8, plugin_agent: *const i8)
 
 /// KCL CLI run function CAPI.
 fn kclvm_cli_run_unsafe(args: *const i8, plugin_agent: *const i8) -> Result<String, String> {
-    let mut args = ExecProgramArgs::from_str(kclvm_runtime::c2str(args));
+    let mut args =
+        ExecProgramArgs::from_str(unsafe { std::ffi::CStr::from_ptr(args) }.to_str().unwrap());
     args.plugin_agent = plugin_agent as u64;
     exec_program(Arc::new(ParseSession::default()), &args)
         .map_err(|e| PanicInfo::from(e).to_json_string())
